biometric_query_system/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pyserial==3.5

